+++ sdk/videoanalyzer/Azure.Media.VideoAnalyzer.Edge/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0-beta.2 (Unreleased)
 
+### Breaking Changes
+- Renamed package from 'Microsoft.Azure.Media.VideoAnalytics.Edge' to 'Azure.Media.VideoAnalyzer.Edge'
+
 
 ## 1.0.0-beta.1 (2020-12-11)
 Initial release